Allowed the detector to ignore small numbers of outliers.
This replaces netevmon's "normal_is_current" variable and its related
algorithm, and is luckily much much simpler. The options may need tuning
and the metric might not be the best, but it does work correctly for
the stated purpose.

It does not correctly ignore two outliers that are separated by one
normal measurement, but other cases appear to work fine.