Fix new edge type to be properly unique by skipping equals and hashCode
Yes, it's highly recommended to implement these, but we don't really
care about the consequences. We just want these edges to all be unique,
and we don't seem to have a good way to make it behave like the
DefaultWeightedEdge which tracks source and target. In fact, I can't
even find where DefaultWeightedEdge defined equals and hashCode...