Prune the graph after some amount of event time has passed.
We prune edges which haven't been seen in a while, and any vertices
that aren't connected to the graph after the edges have been removed.

Event time trigger is not final - it doesn't work well for large
unordered inputs with no watermarks.

We also haven't implemented checkpointing for the prune-related class-
scope variables.